requirements/base.txt
django>=1.9.1
djangorestframework==3.3.2
python-telegram-bot==3.3
Jinja2==2.8
requests==2.9.1